% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_deforestation_rondonia}
\alias{samples_deforestation_rondonia}
\title{Yearly samples of 9 classes on Amazonia biome from Sentinel-2 images}
\format{
A tibble with 6007 samples, with the following columns:
(a) longitude (East-west coordinate of the time series sample in WGS 84),
(b) latitude (North-south coordinate of the time series sample in WGS 84),
(c) start_date (initial date of the time series),
(d) end_date (final date of the time series),
(e) label (the class label associated to the sample),
(f) cube (the name of image collection associated with the data), and
(g) time_series (list containing a tibble with values of the time series).

The samples contain time series for 6007 selected locations which
are assigned to one of the classes "Clear_Cut_Bare_Soil",
"Clear_Cut_Burned_Area", "Mountainside_Forest",
"Forest", "Riparian_Forest", "Clear_Cut_Vegetation",
"Water", "Wetland", and "Seasonally_Flooded".
 Each time series contains values from Sentinel-2/2A
 bands B02, B03, B04, B05, B07, B8A, B08, B11, B12
 from 2022-01-05 to 2022-12-23 in 16-day intervals.
}
\usage{
data("samples_deforestation_rondonia")
}
\description{
A dataset containing a tibble with time series samples from Brazilian
Amazonia biome in the Brazilian state of Rondonia to be used
for deforestation detection.
}
\keyword{datasets}
